requirements.txt
backcall==0.1.0
bleach==3.0.2
blosc==1.8.1
bloscpack==0.16.0
certifi==2018.11.29
chardet==3.0.4
cloudpickle==0.6.1
cycler==0.10.0
dask==1.1.0
decorator==4.3.0
Deprecated==1.2.5
entrypoints==0.2.3
idna==2.8
imageio==2.4.1
ipykernel==5.1.0
ipython==7.2.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.2
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
kiwisolver==1.0.1
lightgbm==2.2.2
line-profiler==2.1.2
MarkupSafe==1.1.0
matplotlib==3.0.2
mistune==0.8.4
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.2
notebook==5.7.4
numpy==1.15.4
opencv-python==4.0.0.21
pandas==0.23.4
pandocfilters==1.4.2
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.4.1
plotly==3.6.0
prometheus-client==0.5.0
prompt-toolkit==2.0.7
ptyprocess==0.6.0
Pygments==2.3.1
pyparsing==2.3.0
python-dateutil==2.7.5
pytz==2018.9
PyWavelets==1.0.1
pyzmq==17.1.2
qtconsole==4.4.3
requests==2.21.0
retrying==1.3.3
scikit-image==0.14.2
scikit-learn==0.20.2
scipy==1.2.0
Send2Trash==1.5.0
Shapely==1.6.4.post2
six==1.12.0
terminado==0.8.1
testpath==0.4.2
toolz==0.9.0
torch==1.0.0
torchfile==0.1.0
torchvision==0.2.1
tornado==5.1.1
tqdm==4.29.0
traitlets==4.3.2
urllib3==1.24.1
visdom==0.1.8.8
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.54.0
widgetsnbextension==3.4.2
wrapt==1.11.1
